// src/resources/tags.ts
export enum TagName {
  Tutorial = 'tutorial',
  React = 'react',
  ThreeJS = 'threejs',
  Particles = 'particles',
  NextJS = 'nextjs',
  GSAP = 'gsap',
  Tailwind = 'tailwind',
  Typescript = 'typescript',
  Responsive = 'responsive',
  PostProcessing = 'post processing',
  VertexShader = 'vertex shader',
  FragmentShader = 'fragment shader',
  Noise = 'noise',
  Startups = 'startups',
  ProductDevelopment = 'product development',
}